+++ tests/user_examples/test_2d_flow_stream_around_fish/2d_flow_stream_around_fish.h
@@ -1,7 +1,6 @@
 
 #include "2d_fish_and_bones.h"
 #include "active_model.h"
-#include "composite_material.h"
 #include "sphinxsys.h"
 
 using namespace SPH;
