--- tests/user_examples/test_2d_flow_stream_around_fish/composite_material.cpp
@@ -1,25 +1,0 @@
-#include "composite_material.h"
-#include "base_local_dynamics.h"
-#include "base_particles.hpp"
-
-namespace SPH
-{
-//=================================================================================================//
-void CompositeSolid::initializeLocalParameters(BaseParticles *base_particles)
-{
-    ElasticSolid::initializeLocalParameters(base_particles);
-    base_particles->registerVariable(material_id_, "MaterialID");
-
-    for (size_t i = 0; i < composite_materials_.size(); ++i)
-    {
-        composite_materials_[i]->initializeLocalParameters(base_particles);
-    }
-}
-//=================================================================================================//
-MaterialIdInitialization::MaterialIdInitialization(SPHBody &sph_body)
-    : LocalDynamics(sph_body), GeneralDataDelegateSimple(sph_body),
-      material_id_(*particles_->getVariableByName<int>("MaterialID")),
-      pos0_(*particles_->getVariableByName<Vecd>("InitialPosition")){};
-//=================================================================================================//
-} // namespace SPH
-  //=================================================================================================//
